+++ app/src/qemu_uart.c
@@ -18,16 +18,12 @@
 /*
 This file implements a simple put char (extended to put string) and use it
 on the _write backend syscall so printf can be used.
-For QEMU machines LM3S6965EVB (Texas Cortex M3) and
-MICROBIT (BBC Cortex-M0).
+For QEMU micro:bit and STM32F030R8 board.
 */
 
-#if (RK_CONF_TRACE == ON)
-#if defined(QEMU_MACHINE_MICROBIT)
+#if ((RK_CONF_TRACE == ON) &&                                                \
+     defined(QEMU_MACHINE_MICROBIT))
 #define TRACE_UART_RX_BUF_SIZE 32U
-#else
-#define TRACE_UART_RX_BUF_SIZE 64U
-#endif
 #define TRACE_UART_RX_BUF_MASK (TRACE_UART_RX_BUF_SIZE - 1U)
 #define NVIC_ISER0 (*(volatile unsigned long *)0xE000E100UL)
 #define NVIC_ICPR0 (*(volatile unsigned long *)0xE000E280UL)
@@ -95,14 +91,97 @@ static void uart_init_once(void)
 }
 #endif
 
-#if defined(QEMU_MACHINE_LM3S6965EVB)
-void kPutc(char const c)
+#if defined(STM32F030x8) || defined(RK_MCU_F030R8)
+#define STM32F030_RCC_BASE (0x40021000UL)
+#define STM32F030_RCC_AHBENR                                                \
+    (*(volatile unsigned long *)(STM32F030_RCC_BASE + 0x14UL))
+#define STM32F030_RCC_APB1ENR                                               \
+    (*(volatile unsigned long *)(STM32F030_RCC_BASE + 0x1CUL))
+#define STM32F030_RCC_AHBENR_GPIOAEN (1UL << 17U)
+#define STM32F030_RCC_APB1ENR_USART2EN (1UL << 17U)
+
+#define STM32F030_GPIOA_BASE (0x48000000UL)
+#define STM32F030_GPIOA_MODER                                               \
+    (*(volatile unsigned long *)(STM32F030_GPIOA_BASE + 0x00UL))
+#define STM32F030_GPIOA_OSPEEDR                                             \
+    (*(volatile unsigned long *)(STM32F030_GPIOA_BASE + 0x08UL))
+#define STM32F030_GPIOA_PUPDR                                               \
+    (*(volatile unsigned long *)(STM32F030_GPIOA_BASE + 0x0CUL))
+#define STM32F030_GPIOA_AFRL                                                \
+    (*(volatile unsigned long *)(STM32F030_GPIOA_BASE + 0x20UL))
+
+#define STM32F030_USART2_BASE (0x40004400UL)
+#define STM32F030_USART2_CR1                                                \
+    (*(volatile unsigned long *)(STM32F030_USART2_BASE + 0x00UL))
+#define STM32F030_USART2_BRR                                                \
+    (*(volatile unsigned long *)(STM32F030_USART2_BASE + 0x0CUL))
+#define STM32F030_USART2_ISR                                                \
+    (*(volatile unsigned long *)(STM32F030_USART2_BASE + 0x1CUL))
+#define STM32F030_USART2_RDR                                                \
+    (*(volatile unsigned long *)(STM32F030_USART2_BASE + 0x24UL))
+#define STM32F030_USART2_TDR                                                \
+    (*(volatile unsigned long *)(STM32F030_USART2_BASE + 0x28UL))
+#define STM32F030_USART_ISR_RXNE (1UL << 5U)
+#define STM32F030_USART_ISR_TXE (1UL << 7U)
+#define STM32F030_USART_CR1_UE (1UL << 0U)
+#define STM32F030_USART_CR1_RE (1UL << 2U)
+#define STM32F030_USART_CR1_TE (1UL << 3U)
+#define STM32F030_USART2_BAUD (115200UL)
+
+static void stm32f030r8_uart_init_once(void)
 {
-    while (UART0_FR & UART0_FR_TXFF)
-        ;
-    UART0_DR = c;
+    static unsigned char init_done;
+    volatile unsigned long fence;
+    unsigned long core_clock;
+
+    if (init_done != 0U)
+    {
+        return;
+    }
+
+    STM32F030_RCC_AHBENR |= STM32F030_RCC_AHBENR_GPIOAEN;
+    STM32F030_RCC_APB1ENR |= STM32F030_RCC_APB1ENR_USART2EN;
+    fence = STM32F030_RCC_AHBENR;
+    fence = STM32F030_RCC_APB1ENR;
+    (void)fence;
+
+    /* NUCLEO-F030R8 exposes USART2 on PA2/PA3 through the ST-LINK VCP. */
+    STM32F030_GPIOA_MODER &= ~((3UL << (2U * 2U)) |
+                               (3UL << (3U * 2U)));
+    STM32F030_GPIOA_MODER |= ((2UL << (2U * 2U)) |
+                              (2UL << (3U * 2U)));
+    STM32F030_GPIOA_AFRL &= ~((0xFUL << (2U * 4U)) |
+                              (0xFUL << (3U * 4U)));
+    STM32F030_GPIOA_AFRL |= ((1UL << (2U * 4U)) |
+                             (1UL << (3U * 4U)));
+    STM32F030_GPIOA_OSPEEDR |= ((3UL << (2U * 2U)) |
+                                (3UL << (3U * 2U)));
+    STM32F030_GPIOA_PUPDR &= ~((3UL << (2U * 2U)) |
+                               (3UL << (3U * 2U)));
+    STM32F030_GPIOA_PUPDR |= (1UL << (3U * 2U));
+
+    core_clock = RK_gSysCoreClock;
+    if (core_clock == 0UL)
+    {
+        core_clock = RK_CONF_SYSCORECLK;
+    }
+    if (core_clock == 0UL)
+    {
+        core_clock = 8000000UL;
+    }
+
+    STM32F030_USART2_CR1 = 0UL;
+    STM32F030_USART2_BRR =
+        (core_clock + (STM32F030_USART2_BAUD / 2UL)) /
+        STM32F030_USART2_BAUD;
+    STM32F030_USART2_CR1 = STM32F030_USART_CR1_UE |
+                           STM32F030_USART_CR1_TE |
+                           STM32F030_USART_CR1_RE;
+    init_done = 1U;
 }
-#elif defined(QEMU_MACHINE_MICROBIT)
+#endif
+
+#if defined(QEMU_MACHINE_MICROBIT)
 void kPutc(char const c)
 {
     uart_init_once();
@@ -113,51 +192,20 @@ void kPutc(char const c)
     UART_EVENTS_TXDRDY = 0;
     UART_TASKS_STOPTX = 1;
 }
+#elif defined(STM32F030x8) || defined(RK_MCU_F030R8)
+void kPutc(char const c)
+{
+    stm32f030r8_uart_init_once();
+    while ((STM32F030_USART2_ISR & STM32F030_USART_ISR_TXE) == 0UL)
+        ;
+    STM32F030_USART2_TDR = (unsigned long)((unsigned char)c);
+}
 #else
 void kPutc(char const c) { (void)c; }
 #endif
 
 #if (RK_CONF_TRACE == ON)
-#if defined(QEMU_MACHINE_LM3S6965EVB)
-int kTraceUartGetc(char *chPtr)
-{
-    if (chPtr == 0)
-    {
-        return (0);
-    }
-
-    if (traceUartRxPop_(chPtr) != 0)
-    {
-        return (1);
-    }
-
-    if (UART0_FR & UART0_FR_RXFE)
-    {
-        return (0);
-    }
-
-    *chPtr = (char)(UART0_DR & 0xFFU);
-    return (1);
-}
-
-void kTraceUartRxEnable(void)
-{
-    UART0_ICR = UART0_RXIC | UART0_RTIC;
-    UART0_IMSC |= UART0_RXIM | UART0_RTIM;
-    traceUartNvicEnable_(UART0_IRQN);
-}
-
-void UART0_Handler(void)
-{
-    while ((UART0_FR & UART0_FR_RXFE) == 0U)
-    {
-        traceUartRxPush_((char)(UART0_DR & 0xFFU));
-    }
-
-    UART0_ICR = UART0_RXIC | UART0_RTIC;
-    kTraceInputSignalFromISR();
-}
-#elif defined(QEMU_MACHINE_MICROBIT)
+#if defined(QEMU_MACHINE_MICROBIT)
 int kTraceUartGetc(char *chPtr)
 {
     if (chPtr == 0)
@@ -197,6 +245,28 @@ void UART0_Handler(void)
         UART_EVENTS_RXDRDY = 0;
         kTraceInputSignalFromISR();
     }
+}
+#elif defined(STM32F030x8) || defined(RK_MCU_F030R8)
+int kTraceUartGetc(char *chPtr)
+{
+    if (chPtr == 0)
+    {
+        return (0);
+    }
+
+    stm32f030r8_uart_init_once();
+    if ((STM32F030_USART2_ISR & STM32F030_USART_ISR_RXNE) == 0UL)
+    {
+        return (0);
+    }
+
+    *chPtr = (char)(STM32F030_USART2_RDR & 0xFFUL);
+    return (1);
+}
+
+void kTraceUartRxEnable(void)
+{
+    stm32f030r8_uart_init_once();
 }
 #else
 int kTraceUartGetc(char *chPtr)